requirements.txt
azure-storage-blob>=12.20.0
